+++ slr_parsing/src/compareexpressions/slr_parsing/tags.py
@@ -1,15 +1,15 @@
 """Helpers for the tags carried by expression-tree nodes."""
 
 
-def tag_rule_union(x, y):
+def union_rule(x, y):
     return x | y
 
 
-def tag_rule_intersection(x, y):
+def intersection_rule(x, y):
     return x & y
 
 
-def tag_transfer(node, rule=tag_rule_union):
+def inherit_tags(node, rule=union_rule):
     tags = set()
     if len(node.children) > 0:
         tags = set(node.children[0].tags)
@@ -18,13 +18,13 @@ def tag_transfer(node, rule=tag_rule_union):
     return tags
 
 
-def tag_removal(node, tag, rule=lambda x: True):
+def remove_tag(node, tag, rule=lambda x: True):
     if rule(node.tags) and tag in node.tags:
         node.tags.remove(tag)
     return node
 
 
-def tag(node, tag=None, rule=lambda x: True):
+def add_tag(node, tag=None, rule=lambda x: True):
     if tag is not None:
         if node.tags is None:
             node.tags = set()
@@ -33,8 +33,8 @@ def tag(node, tag=None, rule=lambda x: True):
     return node
 
 
-def tag_replace(node, old_tag, new_tag):
+def replace_tag(node, old_tag, new_tag):
     if old_tag in node.tags:
-        node = tag(node, new_tag)
-        node = tag_removal(node, old_tag)
+        node = add_tag(node, new_tag)
+        node = remove_tag(node, old_tag)
     return node